Refactor

+++ LunarEngine/Engine/Physics/PhysicsEngine.cs
@@ -11,10 +11,10 @@ namespace LunarEngine.Physics;
 
 public static class PhysicsEngine
 {
-    public const double FIXED_TIMESTAMP = 0.01666667;
-    public static double InterpolatedTime;
+    public const float FIXED_TIMESTAMP = 0.01666667f;
+    public static float InterpolatedTime;
     private static bool _isInPhysicsLoop;
-    public static void TickPhysics(double deltaTime)
+    public static void TickPhysics(float deltaTime)
     {
         _isInPhysicsLoop = true;
         _isInPhysicsLoop = false;
